#112 Add support for uint32 and uint64

+++ option/osenv.go
@@ -88,5 +88,5 @@ func OsLookupEnvUint64(envar string, aDefault uint64) uint64 {
 		panic(err)
 	}
 
-	return uint64(result)
+	return result
 }
